Update game.js
Created 50 unique enemy spawn points so we can make levels where we can control the spot that they spawn in. Fixed the pathing for the hero when he walks off the right side of the screen. He used to transition to the left side of the screen about 200px before hitting the right side. This was because we removed the score panel. I also moved the hero starting point to be at the entrance of the tower. Makes sense right? Then I added kill points that accumulate faster if the hero kills the enemies and slower if the tower kills the enemies.