Game tweaks!
Modified the charging meter so it can accept charging speeds that do not add up to exactly 100. Now we can slowly upgrade the tower's attack speed instead of only choosing from numbers like .1, .5, 1.0, 1.5, 2, 2.5 etc... Now we can do 1.2341234213 if we really want.

Made the range for the tower smaller and more realistic for the starting range.

Made the enemy pathing look better so the enemies no longer line up on the axis for easy kills.
(however, they now move quickly at first and slow down near the tower...)

Made the range for the hero slightly larger as well as the hitbox for the enemies. This basically negates that frustrating feeling that you would get when you change directions just as you think you kill an enemy and it doesn't kill him because the hit box moves to the other side of the hero.

Made the health reduce by 10 for each enemy that hits it.

The game actually feels like a full (and challenging) game now! Woot!